src/data_provider/api/serializers.py
from rest_framework import serializers
from data_provider.models import Event

class EventSerializer(serializers.ModelSerializer):
    class Meta:
        model = Event
        fields = '__all__'
        read_only_fields = ['timestamp']