requirements.txt
Django==4.0.0
djangorestframework==3.13.1
psycopg2-binary==2.9.6
pytest==7.3.0
pytest-django==4.5.2
pytest-mock==3.11.1
pytest-cov==5.0.0
drf-yasg2==1.19.4
celery==5.2.7
redis==4.1.0
django-celery-beat==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1